+++ GNUmakefile
@@ -199,5 +199,5 @@ changelog:
 
 .PHONY: build sweep test testacc fmt fmtcheck lint tools test-compile doc hooks website website-lint website-test
 
-internal-build:
-	./scripts/internal-version-build.sh
+internal-version-build:
+	./scripts/internal-build.sh
